Refactor workflow schemas into separate module
