+++ backend/app/schemas/__init__.py
@@ -1,1 +1,1 @@
-# Schemas Package
+from .workflow import StartWorkflowRequest, CompleteTaskRequest
